+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { Header } from './Header';
+import { Header } from './header/Header';
 import { Content } from './Content';
 import { WipModal } from './WipModal';
